pgtype Float4 and Float8 fields include bit size
e.g. Instead of Float it is Float64. This matches the pattern set by the
database/sql types.

+++ pgtype/builtin_wrappers.go
@@ -279,13 +279,13 @@ func (w *float32Wrapper) ScanFloat64(v Float8) error {
 		return fmt.Errorf("cannot scan NULL into *float32")
 	}
 
-	*w = float32Wrapper(v.Float)
+	*w = float32Wrapper(v.Float64)
 
 	return nil
 }
 
 func (w float32Wrapper) Float64Value() (Float8, error) {
-	return Float8{Float: float64(w), Valid: true}, nil
+	return Float8{Float64: float64(w), Valid: true}, nil
 }
 
 type float64Wrapper float64
@@ -315,13 +315,13 @@ func (w *float64Wrapper) ScanFloat64(v Float8) error {
 		return fmt.Errorf("cannot scan NULL into *float64")
 	}
 
-	*w = float64Wrapper(v.Float)
+	*w = float64Wrapper(v.Float64)
 
 	return nil
 }
 
 func (w float64Wrapper) Float64Value() (Float8, error) {
-	return Float8{Float: float64(w), Valid: true}, nil
+	return Float8{Float64: float64(w), Valid: true}, nil
 }
 
 type stringWrapper string
